+++ src/main/java/com/verdantartifice/primalmagick/common/tiles/rituals/AbstractRitualPropTileEntity.java
@@ -56,14 +56,14 @@ public abstract class AbstractRitualPropTileEntity extends TilePM implements IRi
     }
     
     @Override
-    public CompoundTag save(CompoundTag compound) {
+    protected void saveAdditional(CompoundTag compound) {
+        super.saveAdditional(compound);
         if (this.altarPos != null) {
             compound.putLong("AltarPos", this.altarPos.asLong());
         }
         compound.putBoolean("PropOpen", this.isOpen);
-        return super.save(compound);
     }
-    
+
     @Override
     public void notifyAltarOfPropActivation(float stabilityBonus) {
         if (this.altarPos != null) {